requirements.txt
pyyaml>=6.0
requests>=2.31.0
rich>=13.7.0
prompt_toolkit>=3.0.43
netifaces>=0.11.0

